TP_2/requirements.txt
# Networking Asíncrono
aiohttp

# Parsing HTML
beautifulsoup4
lxml

# Procesamiento de Imágenes
Pillow

# Screenshots y Análisis Avanzado
playwright

# Utilidades Asíncronas
aiofiles   

requests

# Dependencias necesarias solo para el desarrollo y testing
pytest
pytest-asyncio